Moved vscode handlers into separate file src/vscHandlers.ts
